Add Hardcover as another metadata source (#22)
The G——R—— source is private for Reasons™, and we don't currently
include any other implementations for users to actually use or modify.

This adds a new [Hardcover](https://hardcover.app) source, which serves
as both an example and as a concrete alternative for users who'd like to
divorce themselves from G——R——.

Like the original G——R—— source, this new one is also
backwards-compatible with existing libraries because it uses
G——R—— book/work IDs.

Fixes #19.