Load large authors and translated editions asynchronously (#26)
This adds a new `GetAuthorBooks` method to the getter interface which is
responsible for listing the "best" books/editions for an author. When an
author is loaded for the first time, we invoke `GetAuthorBooks`, load
the relevant editions, and denormalize them via `ensureEdition` and
`ensureWork`. Similarly, when a book/edition/translation is loaded for
the first time we also attach it to the work with `ensureEdition`.

Author/work TTLs control how often we refresh, but we actually persist
data for 2x so we can recover previously denormalized data when we
refresh.

Large authors that were already loaded will be incrementally fixed as
works are loaded/refreshed, but I might still want to bust all the
author caches.

In theory this should also fix the translation issue, but in practice
I'm not sure if this is the best UX. R—— has its own caching that might
cause the client hold works that are missing the desired translation. In
the future we might have a "list all editions" method, but I would want
to de-dupe that by title.

Fixes #15
Fixes #17