Disambiguate duplicate titles (#48)
This starts using full titles (e.g. including subtitles) in cases where
an author has several works with the same primary title. We continue to
only use the primary title when it's unambiguous.

I only make this change on the author resource, but technically I should
also update corresponding book/work entries. I'm avoiding that for now
because I want to avoid loops (author updated → work updated → author
updated etc.) and because the author resource tends to be the source of
truth in the UI. This leaves the door open for some inconsistent
behavior when an individual work is refreshed, but it should
self-resolve the next time the author is refreshed.

This fixes https://github.com/blampe/rreading-glasses/issues/35 and
https://github.com/blampe/rreading-glasses/issues/45 but it will take up
to a week to propagate to users. This is because
1. The metadata provider needs to start populating `FullTitle`.
2. Author cache entries to be refreshed.
3. Clients need to refresh author metadata.