Bound async updates (#47)
Right now we eagerly try to sync any new authors we come across, but
these goroutines can pile up and slow each other down because they are
all throttled against the API's RPS limit.

Instead we can offload work to a bounded pool of goroutines (currently
10) to ensure we're always able to make some progress. If we have more
than 10 things to sync they will stay parked in a goroutine until space
opens up in the pool.

Eventually we should record these in the DB so we can recover them on
restart, but this is OK for now.