Correctly de-dupe subtitles when adding works later (#75)
Fixes https://github.com/blampe/rreading-glasses/issues/70

New works can be added to an author after subtitle de-duping has already
taken place. The new work therefore won't get a subtitle applied because
its title already differs, but in some cases it does deserve a subtitle.

Fix this by keeping a record of what the original (short) title was, and
use that (if available) for de-duping.

While we're in here let's also make this case-insensitive.

This will take about a month for this to fully roll out because cache
entries need to expire.