Only compress data at rest (#79)
We don't gain much by storing compressed cache values in memory, in fact
we end up allocating a lot of slices to decompress things. PG hits are
fast enough, so the size of our in-memory set isn't too important. (It's
currently around ~28k items for ~1.8G RAM.)

This change introduces a buffer pool to help reduce GC churn. This is a
sufficient fix, but we also push compression down to only the PG layer.
This will make #28 easier to reason about.