Eventually retry 404s (#105)
We have a bug where if we ever store a `_missing` sentinel we'll never
refresh the cache entry. This fixes that.

We also no longer need to store things with the confusing 2x TTL. That
was to ensure we still had data available after TTL expiration, but
we're always returning stale data now so it's a non-issue.

We also start fuzzing TTLs a bit to spread out load.