Repair broken sort invariants (#106)
A user reported an issue where they weren't able to add an author.
Ultimately it was due to the author having a duplicated work, but this
shouldn't happen under normal circumstances. We keep the author's works
sorted by foreign id, and we rely on this sort order when updating
things.

When I first introduced the sorted invariant I had some logic to sort
things that weren't already sorted. I thought enough time had passed for
it to be safe to remove that logic, but that might not have been the
case. Or there might be a different bug causing things to get out of
order.

In any case, this adds some logic to repair + log when we find things
out of order. It also makes our denormalization steps more efficient.