Return individualized errors when batching queries (#112)
Right now if one of our queries fails, for example with a 404, the
entire batch of queries will get the same error.

This modifies our error handling to inspect the returned `gqlerror.List`
and return errors to the appropriate subscribers.

While testing this I happened to trigger the WAF, so I'm also modifying
the auth logic to fall back to using the default token if we fail to
refresh creds. The alternative will usually put the crash into a crash
loop, which makes the situation worse.