Start automatically populating some editions (#130)
This starts automatically populating up to 20 editions per work. This
happens asynchronously, so the very first time a work is fetched it will
still only have the "best" edition, but subsequent fetches may include
more.

We're limited to 20 right now partly because it remains to be seen what
kind of impact this will have on performance, but more practically
because 20 is the limit before we need to paginate and this is just
easier for now.

Editions are de-duped somewhat, so for example if two editions share the
same language and title we only take one. (The official server's edition
selection menu can have dozens and dozens of duplicated titles which are
a mess to crawl through.)

I realized I'm not returning the ISO639-3 code for language, which might
impact edition selection. This might also explain why users have still
had trouble picking up translated editions.

Implementation-wise I just plumbed a callback through to the getter
which allows it to register some IDs for fetching. Blunt but it works.
We're getting books in serial during denormalization, so there's room
for improvement.

Fixes https://github.com/blampe/rreading-glasses/issues/82 with the
usual caveat that caches will take a while to expire.