Speed up edition loading (#175)
We weren't always loading editions are part of GetBook (only GetWork).
We load authors with GetAuthorBooks so it's important we also grab
editions as part of that.

We were also doing fetching as part of the denorm channel which backs up
everything significantly. I've tried to move all the actual fetching
into the refresh group so the denorm can go back to only handling
serialization.